requirements.txt
paho-mqtt==1.6.1
requests==2.25.1
pandas==1.2.4
jsonschema==3.2.0
dos2unix=1.2.1